+++ changelog.md
@@ -13,6 +13,7 @@
 
 
 * [Changelog](#changelog)
+  * [Releases v1.8.6](#releases-v186)
   * [Releases v1.8.5](#releases-v185)
   * [Releases v1.8.4](#releases-v184)
   * [Releases v1.8.3](#releases-v183)
@@ -45,11 +46,15 @@
 
 ## Changelog
 
+### Releases v1.8.6
+
+1. Fix bug not supporting boards. Check [Missing ESP definitions ..... #45](https://github.com/khoih-prog/EthernetWebServer/issues/45)
+
 ### Releases v1.8.5
 
 1. Restore support to AVR Mega2560
 2. Add support to megaAVR boards such as `Arduino Nano Every`, `UNO WiFi Rev 2` using `ATmega4809` MPU
-2. Fix libb64 fallthrough compile warning
+3. Fix libb64 fallthrough compile warning
 
 ### Releases v1.8.4